Set dummy password and refactoring
